'10_jpa_repo_update_done'

+++ 12_spring_data_jpa_repo_getting_start/src/main/java/com/jpa/repositories/StateRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.jpa.entities.State;
 
-public interface StateRepo extends JpaRepository<State, Long> {
+public interface StateRepo extends JpaRepository<State, Integer> {
 
 }
